requirements.txt
streamlit
mysql-connector-python
pandas
streamlit_option_menu
streamlit_ace
matplotlib